pyproject.toml
[tool.poetry]
name = "instagraper"
version = "0.1.0"
description = ""
authors = ["Francisco Macedo <franciscovcbm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-decouple = "^3.8"
typer = {extras = ["all"], version = "^0.9.0", optional = true}


[tool.poetry.scripts]
instagraper = "instagraper.cli:app"


[tool.poetry.extras]
cli = ["typer"]


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.8"
requests-mock = "^1.11.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
